refactored a bit and move out of wrappers and into the base utils
